+++ edsnlp/data/converters.py
@@ -19,7 +19,6 @@ from typing import (
     Union,
 )
 
-import pydantic
 import spacy
 from confit.registry import ValidatedFunction
 from spacy.tokenizer import Tokenizer
@@ -90,9 +89,7 @@ def validate_kwargs(func, kwargs):
         model = vd.init_model_instance(
             **{k: v for k, v in kwargs.items() if k in spec.args}
         )
-        fields = (
-            model.__fields__ if pydantic.__version__ < "2" else vd.model.model_fields
-        )
+        fields = vd.model.model_fields
         d = {
             k: v
             for k, v in model.__dict__.items()